% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mst.R
\name{adj_matrix_to_edges}
\alias{adj_matrix_to_edges}
\title{Convert an adjacency matrix into a list of edges (source, destination, weight)}
\usage{
adj_matrix_to_edges(adj_matrix)

adj_matrix_to_edges(adj_matrix)
}
\arguments{
\item{adj_matrix}{Weighted adjacency matrix}
}
\value{
Matrix of edges (source, destination, weight)

Matrix of edges (source, destination, weight)
}
\description{
Convert an adjacency matrix into a list of edges (source, destination, weight)

Convert an adjacency matrix into a matrix of edges (source, destination, weight)
}
