% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mst.R
\name{generate_random_adjacency_matrix}
\alias{generate_random_adjacency_matrix}
\title{Generate weighted adjacency matrix of a graph with given number of vertices}
\usage{
generate_random_adjacency_matrix(n, num_zeros = NULL)

generate_random_adjacency_matrix(n, num_zeros = NULL)
}
\arguments{
\item{n}{Number of vertices}

\item{num_zeros}{Number of zeros to add randomly}
}
\value{
NumericMatrix representing the weighted adjacency matrix

A weighted adjacency matrix representing the random graph
}
\description{
Generate weighted adjacency matrix of a graph with given number of vertices
}
