% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{kruskal_mst}
\alias{kruskal_mst}
\title{Find the Minimum Spanning Tree using Kruskal's algorithm}
\usage{
kruskal_mst(adj_matrix)
}
\arguments{
\item{adj_matrix}{Weighted adjacency matrix}
}
\value{
Weighted adjacency matrix of the Minimum Spanning Tree
}
\description{
Find the Minimum Spanning Tree using Kruskal's algorithm
}
