% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/mst.R
\name{min_key_vertex}
\alias{min_key_vertex}
\title{Find the vertex with minimum key value that is not yet in the MST}
\usage{
min_key_vertex(key, mst_set)

min_key_vertex(key, mst_set)
}
\arguments{
\item{key}{Key values for vertices}

\item{mst_set}{A logical vector indicating whether each vertex is in the MST}
}
\value{
Index of the vertex with the minimum key value

Index of the vertex with the minimum key value that is not yet in the MST
}
\description{
Find the vertex with minimum key value that is not yet in the MST

Find the vertex with minimum key value that is not yet in the MST
}
