% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prim_mst_rcpp}
\alias{prim_mst_rcpp}
\title{Find the Minimum Spanning Tree using Prim's algorithm}
\usage{
prim_mst_rcpp(adj_matrix)
}
\arguments{
\item{adj_matrix}{Weighted adjacency matrix}
}
\value{
Weighted adjacency matrix of the Minimum Spanning Tree
}
\description{
Find the Minimum Spanning Tree using Prim's algorithm
}
